Fix signed/unsigned loop
players->sBuf is an int32, but when it is being read, it has no way of
knowing that it was actually an unsigned int, so it needed to be cast

+++ src/client/clientControllerState.cpp
@@ -252,7 +252,7 @@
 
             data->model->addData( "voteState", false );
 
-            for ( unsigned int i = 0; i < sBuf->players( ); i++ ) {
+            for ( unsigned int i = 0; i < ( unsigned int )sBuf->players( ); i++ ) {
                 data->model->addData( std::string( "player" ) + std::to_string( i ), NULL );
             }
